Complete transaction before sleeping for polling interval
The subscriber query method is responsible for sleeping for the polling interval
when there are no remaining messages. It relies on the `defer`'d cleanup to
either commit or rollback the transaction.

In production deployments the polling interval may be greater than the
transaction idle limit, sometimes significantly so.

In that situation, the transaction will be terminated by the database server
before `query` returns, thereby causing the commit to fail.

This commit smooths this over by explicitly rolling back the transaction when
there are no messages to consume.